// src/app/utils/models/Register-user.model.ts
export class RegisterUserModel{
    cardType: any;
    prefix:any;
    cardNumber: any
    firstName: any;
    lastName: any;
    gender: any;
    dateOfBirth: any;
    nationality: any;
    mobileNo: any;
    landlineNo: any;
    email: any;
    address: any;
    Country: any;
    street: any;
    city: any;
    area: any;
    houseNo: any;
    Locale: any;
    ExtraMilesCheck: any;
}